pyproject.toml
[project]
name = "ai-code-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.118.3",
    "google-adk==1.19.0",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.0",
    "litellm>=1.78.0",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "python-multipart==0.0.20",
    "runtime>=0.1.4",
    "starlette==0.48.0",
    "uvicorn==0.38.0",
    "streamlit>=1.52.1",
]
